refactored code again, moved exceptions in service, finished unit tests

+++ src/main/java/com/cognizant/practice/blog/article/service/ArticlesService.java
@@ -1,5 +1,6 @@
 package com.cognizant.practice.blog.article.service;
 
+import com.cognizant.practice.blog.article.convertor.ArticleConvertor;
 import com.cognizant.practice.blog.article.dto.Article;
 import com.cognizant.practice.blog.article.entity.ArticleEntity;
 import com.cognizant.practice.blog.article.dto.ArticleRequest;
@@ -26,43 +27,59 @@ public class ArticlesService {
         this.articleRepository = articlesRepository;
     }
 
+    public boolean isValidParam(String param) {
+        return !StringUtils.isEmpty(param);
+    }
+
+    public boolean isValidRequest(ArticleRequest request) {
+        return isValidParam(request.title()) && isValidParam(request.content());
+    }
+
     public List<Article> getAllArticles() {
         List<ArticleEntity> entities = articleRepository.findAll();
 
-        return entities.stream().map(articleEntity -> new Article(
-                articleEntity.getId(),
-                articleEntity.getTitle(),
-                articleEntity.getContent(),
-                articleEntity.getCreatedDate(),
-                articleEntity.getUpdatedDate()
-        )).collect(Collectors.toList());
+        return entities.stream().map(ArticleConvertor::toDto).collect(Collectors.toList());
     }
 
-    public Optional<Article> getArticleById(UUID id) {
+    public Article getArticleById(UUID id) {
         Optional<ArticleEntity> article = articleRepository.findById(id);
 
-        return article.map(articleEntity -> new Article(
-                articleEntity.getId(),
-                articleEntity.getTitle(),
-                articleEntity.getContent(),
-                articleEntity.getCreatedDate(),
-                articleEntity.getUpdatedDate()
-        ));
+        if(article.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Article not found");
+        }
+
+        return ArticleConvertor.toDto(article.get());
     }
 
     public void deleteArticle(UUID id) {
+        Optional<ArticleEntity> article = articleRepository.findById(id);
+        if (article.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Article not found");
+        }
+
         articleRepository.deleteById(id);
     }
 
     public Article createArticle(ArticleRequest articleRequest) {
-        ArticleEntity newArticle = new ArticleEntity(null, articleRequest.title(), articleRequest.content(), LocalDateTime.now(), LocalDateTime.now(), null);
-        articleRepository.save(newArticle);
+        if (!isValidRequest(articleRequest)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Fields can not be empty");
+        }
 
-        return new Article(newArticle.getId(), newArticle.getTitle(), newArticle.getContent(), newArticle.getCreatedDate(), newArticle.getUpdatedDate());
+        ArticleEntity newArticle = new ArticleEntity(null, articleRequest.title(), articleRequest.content(), LocalDateTime.now(), LocalDateTime.now(), null);
+
+        return ArticleConvertor.toDto(articleRepository.save(newArticle));
     }
 
     public Article updateArticle(UUID id, ArticleRequest articleRequest) {
+        if (!isValidRequest(articleRequest)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Fields can not be empty");
+        }
+
         Optional<ArticleEntity> article = articleRepository.findById(id);
+
+        if(article.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Article not found");
+        }
 
         ArticleEntity newArticle = article.get();
 
@@ -70,8 +87,7 @@ public class ArticlesService {
         newArticle.setTitle(articleRequest.title());
         newArticle.setUpdatedDate(LocalDateTime.now());
 
-        articleRepository.save(newArticle);
-        return new Article(newArticle.getId(), newArticle.getTitle(), newArticle.getContent(), newArticle.getCreatedDate(), newArticle.getUpdatedDate());
+        return ArticleConvertor.toDto(articleRepository.save(newArticle));
     }
 
 }
